+++ NEWS.md
@@ -1,3 +1,8 @@
+# encore.analytics 0.3.0
+
+* `simulate_data()` does not distribute missingness uniformly across all variables anymore, but rather introduces a different (random) missingness on different columns which average to the specified proportion missingness.
+* Added function to generate study design diagrams....
+
 # encore.analytics 0.2.0
 
 * Initial version of the package